+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -334,7 +334,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KVCH33GP7ECJ8FABA3CHFD14"
+        "01KVFA5MARX9EYQ4H5V860CP1E"
     )
     r"""The unique identifier of the data source"""