+++ packages/orq-rc/src/orq_ai_sdk/models/post_v2_human_eval_setsop.py
@@ -188,7 +188,7 @@ class PostV2HumanEvalSetsResponseBody2(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-06-18T04:50:35.672Z")
+    updated: Optional[datetime] = parse_datetime("2026-06-19T06:47:32.188Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")
@@ -273,7 +273,7 @@ class PostV2HumanEvalSetsResponseBody1(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-06-18T04:50:35.672Z")
+    updated: Optional[datetime] = parse_datetime("2026-06-19T06:47:32.188Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")