+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasetop.py
@@ -160,7 +160,7 @@ class UpdateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-06-18T04:50:35.672Z")
+    updated: Optional[datetime] = parse_datetime("2026-06-19T06:47:32.188Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")