+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KVCH33GT816150N4FDESP2ZH"
+        "01KVFA5MAV9H8M0D21WWRNVPFH"
     )
     r"""The unique identifier of the data source"""